feat: [CC-521] Add video frame adjustments for cover view
