+++ Source/Filters/Video/YPVideoFiltersVC.swift
@@ -85,8 +85,8 @@ open class YPVideoFiltersVC: UIViewController, IsMediaFilterVC {
         let v = YPVideoView()
         return v
     }()
-    public var coverImageView: UIImageView = {
-        let v = UIImageView()
+    public var coverImageView: YPCoverImageView = {
+        let v = YPCoverImageView()
         v.contentMode = .scaleAspectFit
         v.isHidden = true
         return v
@@ -104,6 +104,8 @@ open class YPVideoFiltersVC: UIViewController, IsMediaFilterVC {
 
         videoView.cropRect = inputVideo.cropRect // pass the crop rect over to the video so it can present the video relative to the crop rect
         videoView.asset = inputVideo.asset // pass the asset over so we can use it to determine the original video dimensions
+        coverImageView.cropRect = inputVideo.cropRect // pass the crop rect over to the video so it can present the video relative to the crop rect
+        coverImageView.asset = inputVideo.asset // pass the asset over so we can use it to determine the original video dimensions
 
         mediaManager.initialize()
         
@@ -124,6 +126,7 @@ open class YPVideoFiltersVC: UIViewController, IsMediaFilterVC {
                          object: videoView.player.currentItem)
 
         videoView.clipsToBounds = true
+        coverImageView.clipsToBounds = true
 
         // configure progress view
         view.addSubview(progressView)
