+++ Source/Filters/Video/YPVideoView.swift
@@ -13,14 +13,14 @@ import Photos
 
 /// A video view that contains video layer, supports play, pause and other actions.
 /// Supports xib initialization.
-public class YPVideoView: UIView {
+public class YPVideoView: YPAdjustableView {
     public let playImageView = UIImageView(image: nil)
 
     internal var playerView = YPVideoPlayer()
     internal var previewImageView = UIImageView()
 
     var cropRect: CGRect?
-    var asset: PHAsset?
+    public var asset: PHAsset?
 
     public var targetAspectRatio: CGFloat?
 
@@ -45,16 +45,11 @@ public class YPVideoView: UIView {
     public override func layoutSubviews() {
         super.layoutSubviews()
 
+        updateViewFrameAction = updateViewFrame
+
         // Ensure necessary properties are available
         guard let cropRect = cropRect, let asset = asset else { return }
-        let assetSize = CGSize(width: CGFloat(asset.pixelWidth), height: CGFloat(asset.pixelHeight))
-
-        // Check if the aspect ratio adjustment is needed and call appropriate methods
-        if let aspectRatio = targetAspectRatio, aspectRatio != assetSize.width / assetSize.height {
-            adjustPlayerViewFrameForAspectRatio(cropRect: cropRect, aspectRatio: aspectRatio, assetSize: assetSize)
-        } else {
-            adjustPlayerViewFrame(cropRect: cropRect, assetSize: assetSize)
-        }
+        adjustViewFramesIfNeeded(cropRect: cropRect, asset: asset, targetAspectRatio: targetAspectRatio)
     }
 
     internal func setup() {
@@ -96,70 +91,8 @@ public class YPVideoView: UIView {
             player.play()
         }
     }
-}
 
-// MARK: - Player View Adjustment
-extension YPVideoView {
-    // Method to adjust the player view frame considering the asset's original size
-    private func adjustPlayerViewFrame(cropRect: CGRect, assetSize: CGSize) {
-        let assetRect = CGRect(origin: .zero, size: assetSize)
-        let fitFrame = calculateAspectFitFrame(assetSize: assetSize)
-        let targetFrame = calculateTargetFrame(cropRect: cropRect, assetRect: assetRect, fitFrame: fitFrame)
-        updatePlayerViewFrame(targetFrame)
-    }
-
-    // Method to adjust the player view frame based on a target aspect ratio
-    private func adjustPlayerViewFrameForAspectRatio(cropRect: CGRect, aspectRatio: CGFloat, assetSize: CGSize) {
-        let adjustedAssetSize = adjustedSizeForAspectRatio(assetSize, aspectRatio: aspectRatio)
-        let fitFrame = calculateAspectFitFrame(assetSize: adjustedAssetSize)
-        let targetFrame = calculateTargetFrame(cropRect: cropRect, assetRect: CGRect(origin: .zero, size: adjustedAssetSize), fitFrame: fitFrame, adjustForAspectRatio: true)
-        updatePlayerViewFrame(targetFrame)
-    }
-
-    // Calculate the adjusted size for the asset based on the target aspect ratio
-    private func adjustedSizeForAspectRatio(_ assetSize: CGSize, aspectRatio: CGFloat) -> CGSize {
-        // Adjust the asset size to maintain the target aspect ratio
-        if assetSize.width / assetSize.height > aspectRatio {
-            return CGSize(width: assetSize.height * aspectRatio, height: assetSize.height)
-        } else {
-            return CGSize(width: assetSize.width, height: assetSize.width / aspectRatio)
-        }
-    }
-
-    // Calculate the frame for aspect fit scaling of the asset
-    private func calculateAspectFitFrame(assetSize: CGSize) -> CGRect {
-        // Determine the scale factor to fit the asset in the current bounds
-        let widthScale = bounds.size.width / assetSize.width
-        let heightScale = bounds.size.height / assetSize.height
-        let scale = min(widthScale, heightScale)
-        return CGRect(origin: .zero, size: CGSize(width: assetSize.width * scale, height: assetSize.height * scale))
-    }
-
-    // Calculate the target frame based on the crop rectangle and asset size
-    private func calculateTargetFrame(cropRect: CGRect, assetRect: CGRect, fitFrame: CGRect, adjustForAspectRatio: Bool = false) -> CGRect {
-        // Determine the scaling factor needed for cropping
-        let scale = max(assetRect.size.width / cropRect.size.width, assetRect.size.height / cropRect.size.height)
-        var targetFrame = fitFrame.applying(CGAffineTransform(scaleX: scale, y: scale))
-
-        // Calculate offset scaling relative to the crop rectangle and asset rectangle centers
-        let xOffsetScale = (cropRect.midX - assetRect.midX) / assetRect.size.width
-        let yOffsetScale = (cropRect.midY - assetRect.midY) / assetRect.size.height
-
-        // Adjust the target frame origin based on the calculated offsets
-        if adjustForAspectRatio {
-            targetFrame.origin.x = 0.5 * (bounds.size.width - targetFrame.size.width) - xOffsetScale
-            targetFrame.origin.y = 0.5 * (bounds.size.height - targetFrame.size.height) - yOffsetScale
-        } else {
-            targetFrame.origin.x = 0.5 * (bounds.size.width - targetFrame.size.width) - xOffsetScale * targetFrame.size.width
-            targetFrame.origin.y = 0.5 * (bounds.size.height - targetFrame.size.height) - yOffsetScale * targetFrame.size.height
-        }
-
-        return targetFrame
-    }
-
-    // Update the frame and video gravity of the player view
-    private func updatePlayerViewFrame(_ frame: CGRect) {
-        // Set the new frame to the player view layer and adjust video gravity
+    private func updateViewFrame(_ frame: CGRect) {
         playerView.layer.frame = frame
         playerView.playerLayer.videoGravity = .resizeAspectFill
     }
