+++ Source/Resources/en.lproj/YPImagePickerLocalizable.strings
